Vite
Looking for the best way to host my frontend app I decided that the best
 way to move forward is to use vite instead of hosting it myself. Beside
optimization, it botheres me that the assesment being stictly frontend had
a server side to it. So I decided that it is better to use a development
server provided by vite and take the time to truly work only on the
frontend side of things
Deleted the dist folder since vite handles compiling my typescript,
one more reason it was decided to use it.
Moved the index.html to the root folder
Moved the styles folder on the public folder as suggested by vite docs