Input event
Created events.ts file where the plan is to store various events
Created event listener for the input component where when the user types the movie they are
looking for the app fetches movies from the tmdb api according to the input value.
If the input is empty, the app renders the starter page.
Functionality is at this point not complete.