url router
The router, now with the use of global variables and the router object is capable of handling all the wanted requests
and update the html accordingly while any attempt to navigate to a different route than the one specifed
on the url object, will navigate the user to the 404 page. Additionally now typing on the input redirects the user
to /search while deleteing the input value will redirect them to the / page. Finally being on the /search
 page and reloading the page or manually requesting it will also redirect them to the home page